Add Krylov support
